requirements.txt
certifi==2018.4.16
chardet==3.0.4
idna==2.6
libusb1==1.6.4
ndeflib==0.3.2
nfcpy==0.13.4
pyDes==2.0.1
pyserial==3.4
requests==2.18.4
soco==0.14
urllib3==1.22
xmltodict==0.11.0
